#12. Add PostService with filling a post instance
